Refactored to remove telemetry handling to its own service
